fix: volta metadata
